backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
easyocr==1.7.1
openai==1.40.0
google-generativeai==0.7.2
youtube-transcript-api==0.6.2
Pillow==10.3.0
